requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brotli-bin==1.0.9
brotli==1.0.9
bzip2==1.0.8
c-ares==1.19.1
ca-certificates==2023.5.7
cached-property==1.5.2
cached_property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.2
flask==2.3.2
flatbuffers==23.3.3
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.3.3
gast==0.4.0
giflib==5.2.1
google-auth-oauthlib==0.4.6
google-auth==2.17.3
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.9.0
hdf5==1.14.0
icu==72.1
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.23.2
ipython-genutils==0.2.0
ipython==8.14.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter-client==8.2.0
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.6.3
jupyter-server-terminals==0.4.4
jupyter-server==2.6.0
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras-preprocessing==1.1.2
keras==2.12.0
kiwisolver==1.4.4
krb5==1.20.1
leven==1.0.4
libabseil==20220623.0
libaec==1.0.6
libblas==3.9.0
libbrotlicommon==1.0.9
libbrotlidec==1.0.9
libbrotlienc==1.0.9
libcblas==3.9.0
libcurl==8.1.2
libcxx==16.0.6
libedit==3.1.20191231
libev==4.33
libffi==3.4.2
libgfortran5==12.2.0
libgfortran==5.0.0
libgrpc==1.51.1
libjpeg-turbo==2.1.5.1
liblapack==3.9.0
libnghttp2==1.52.0
libopenblas==0.3.23
libpng==1.6.39
libprotobuf==3.21.12
libsqlite==3.42.0
libssh2==1.11.0
libzlib==1.2.13
llvm-openmp==16.0.6
markdown==3.4.3
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.1
mediapipe==0.10.1
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
ncurses==6.4
nest-asyncio==1.5.6
networkx==3.1
nose==1.3.7
notebook-shim==0.2.3
notebook==6.5.4
numpy==1.25.0
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
openssl==3.1.1
opt_einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
pip==23.1.2
platformdirs==3.7.0
pooch==1.7.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1-modules==0.2.7
pyasn1==0.4.8
pycparser==2.21
pygments==2.15.1
pyjwt==2.7.0
pyopenssl==23.2.0
pyparsing==3.1.0
pyrsistent==0.19.3
pysocks==1.7.1
python-dateutil==2.8.2
python-flatbuffers==23.5.26
python-json-logger==2.0.7
python==3.10.11
python_abi==3.10
pytz==2023.3
pyu2f==0.1.5
pyyaml==6.0
pyzmq==25.1.0
qtconsole==5.4.3
qtpy==2.3.1
re2==2023.02.01
readline==8.2
requests-oauthlib==1.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
send2trash==1.8.2
setuptools==67.7.2
six==1.16.0
sklearn==0.0.post5
snappy==1.1.10
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.11.2
tensorflow-base==2.11.1
tensorflow-estimator==2.11.1
tensorflow==2.11.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==8.6.12
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.6.3
typing_extensions==4.6.3
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.0
werkzeug==2.3.6
wheel==0.40.0
widgetsnbextension==4.0.7
wrapt==1.15.0
xz==5.2.6
yarl==1.9.2
zipp==3.15.0
zlib==1.2.13
zstd==1.5.2
